refactor(db): split campaigns.ts by responsibility (refs #814)
Break the 835-line campaigns.ts monolith into three focused modules:
- campaigns/types.ts   — types, constants, validation helpers, row mappers
- campaigns/crud.ts    — campaign CRUD + active-engagement cache read
- campaigns/sends.ts   — send lifecycle: create, claim, ack, stats
- campaigns/index.ts   — barrel re-export (all 22 exports re-surfaced)
- campaigns.ts         — thin delegation shim (original path unchanged)

Zero behavior change. All 7738 tests pass; typecheck and lint clean.
No import sites updated — existing @/lib/db/campaigns path continues to
resolve through the shim barrel.
