+++ apps/web/lib/db/campaigns/index.ts
@@ -1,0 +1,44 @@
+/**
+ * Barrel re-export for the campaigns DB layer.
+ *
+ * All external importers use `@/lib/db/campaigns` — this index keeps that
+ * path stable while the implementation is split into cohesive sub-modules:
+ *
+ *   types.ts  — shared types, constants, validation, row mappers
+ *   crud.ts   — campaign CRUD (get/create/update/delete + active engagement)
+ *   sends.ts  — send lifecycle (create, claim, ack, stats)
+ */
+
+export type {
+  CampaignType,
+  CampaignStatus,
+  CampaignSendStatus,
+  Campaign,
+  CampaignSend,
+  CampaignSendStats,
+} from "./types";
+
+export {
+  CampaignRowSchema,
+  CampaignSendRowSchema,
+  mapCampaignRow,
+  mapSendRow,
+} from "./types";
+
+export {
+  dbGetCampaigns,
+  dbGetCampaign,
+  dbCreateCampaign,
+  dbUpdateCampaign,
+  dbDeleteCampaign,
+  dbGetActiveEngagementCampaign,
+} from "./crud";
+
+export {
+  dbCreateCampaignSends,
+  dbGetPendingSends,
+  dbClaimPendingSends,
+  dbMarkSendsSent,
+  dbMarkSendsFailed,
+  dbGetCampaignStats,
+} from "./sends";
